Fixed wrong checking order on guildmemberadd.ts
