+++ src/Client/index.ts
@@ -9,8 +9,7 @@ import {
 import { connect } from 'mongoose';
 import path from 'path';
 import { readdirSync } from 'fs';
-import { Command, Config, SlashCommand } from '../Interfaces';
-import ConfigJson from '../config.json';
+import { Command, SlashCommand } from '../Interfaces';
 import dotenv from 'dotenv';
 import { REST } from '@discordjs/rest';
 import { Routes } from 'discord-api-types/v9';
@@ -27,7 +26,6 @@ class ExtendedClient extends Client {
   public events: Collection<string, Command> = new Collection();
   public cooldowns: Collection<string, Collection<string, number>> = new Collection();
   public categories: Set<string> = new Set();
-  public config: Config = ConfigJson;
   public aliases: Collection<string, Command> = new Collection();
   public embed(options: MessageEmbedOptions, message: Message): MessageEmbedOptions {
     return new MessageEmbed({ ...options })
