--- src/Interfaces/Config.ts
@@ -1,3 +1,0 @@
-export interface Config {
-    prefix: String;
-}
