+++ src/Interfaces/index.ts
@@ -1,4 +1,3 @@
 export { Command } from './Command';
 export { Event } from './Event';
-export { Config } from './Config';
 export { SlashCommand } from './SlashCommand';
